Simple Inference added.
