+++ docs/source/ttnn/ttnn/tutorials.rst
@@ -39,3 +39,4 @@ Available tutorials:
    tutorials/2025_dx_rework/ttnn_mlp_inference_mnist.ipynb
    tutorials/2025_dx_rework/ttnn_multihead_attention.ipynb
    tutorials/2025_dx_rework/ttnn_basic_conv.ipynb
+   tutorials/2025_dx_rework/ttnn_simplecnn_inference.ipynb
